+++ CHANGELOG.md
@@ -4,9 +4,10 @@ Change history of the project. All the feature updates, bug fixes, breaking chan
 
 ## [ 0.2.0 ]
 
-- Fields chooser with support for multiple tables.
-- Table chooser with support for all available tables.
+- Fields chooser with support for multiple tables
+- Table chooser with support for all available tables
 - Template variables list support
+- Custom description support in annotation
 
 ## [ 0.1.0 ]